api-discovery/requirements.txt
# Core libraries for API Discovery with ChromaDB and OpenAI
chromadb==0.4.22
openai==1.12.0
langchain==0.1.0
langchain-openai==0.0.5

# Data processing and utilities
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.2
markdown==3.5.2
pyyaml==6.0.1
lxml==4.9.3

# Jupyter and visualization
jupyter==1.0.0
ipykernel==6.27.1
matplotlib==3.7.2
seaborn==0.12.2

# Text processing and embeddings
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.2

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
